// healthmate-backend-public/src/shared/types/jwt.type.ts
export interface AccessTokenPayloadCreate {
  userId: string;
  deviceId: string;
  roleId: string;
  roleName: string;
}

export interface AccessTokenPayload extends AccessTokenPayloadCreate {
  exp: number;
  iat: number;
}

export interface RefreshTokenPayloadCreate {
  userId: string;
}

export interface RefreshTokenPayload extends RefreshTokenPayloadCreate {
  exp: number;
  iat: number;
}